Move test_list_objects to test_s3_objects_operations.py
